feat(db): extract model define and add auto create table func
