+++ pkg/cluster/ClusterRepository.go
@@ -17,41 +17,11 @@
 package repository
 
 import (
-	"github.com/devtron-labs/kubelink/pkg/remoteConnection"
-	"github.com/devtron-labs/kubelink/pkg/sql"
+	"github.com/devtron-labs/kubelink/pkg/model"
 	"github.com/go-pg/pg"
 	"github.com/go-pg/pg/orm"
 	"go.uber.org/zap"
 )
-
-type Cluster struct {
-	tableName                struct{}          `sql:"cluster" pg:",discard_unknown_columns"`
-	Id                       int               `sql:"id,pk"`
-	ClusterName              string            `sql:"cluster_name"`
-	ServerUrl                string            `sql:"server_url"`
-	RemoteConnectionConfigId int               `sql:"remote_connection_config_id"`
-	PrometheusEndpoint       string            `sql:"prometheus_endpoint"`
-	Active                   bool              `sql:"active,notnull"`
-	CdArgoSetup              bool              `sql:"cd_argo_setup,notnull"`
-	Config                   map[string]string `sql:"config"`
-	PUserName                string            `sql:"p_username"`
-	PPassword                string            `sql:"p_password"`
-	PTlsClientCert           string            `sql:"p_tls_client_cert"`
-	PTlsClientKey            string            `sql:"p_tls_client_key"`
-	AgentInstallationStage   int               `sql:"agent_installation_stage"`
-	K8sVersion               string            `sql:"k8s_version"`
-	ErrorInConnecting        string            `sql:"error_in_connecting"`
-	IsVirtualCluster         bool              `sql:"is_virtual_cluster"`
-	InsecureSkipTlsVerify    bool              `sql:"insecure_skip_tls_verify"`
-	ProxyUrl                 string            `sql:"proxy_url"`
-	ToConnectWithSSHTunnel   bool              `sql:"to_connect_with_ssh_tunnel"`
-	SSHTunnelUser            string            `sql:"ssh_tunnel_user"`
-	SSHTunnelPassword        string            `sql:"ssh_tunnel_password"`
-	SSHTunnelAuthKey         string            `sql:"ssh_tunnel_auth_key"`
-	SSHTunnelServerAddress   string            `sql:"ssh_tunnel_server_address"`
-	RemoteConnectionConfig   *remoteConnection.RemoteConnectionConfig
-	sql.AuditLog
-}
 
 func NewClusterRepositoryImpl(dbConnection *pg.DB, logger *zap.SugaredLogger) *ClusterRepositoryImpl {
 	return &ClusterRepositoryImpl{
@@ -66,13 +36,13 @@ type ClusterRepositoryImpl struct {
 }
 
 type ClusterRepository interface {
-	FindAllActive() ([]*Cluster, error)
-	FindById(id int) (*Cluster, error)
-	FindByIdWithActiveFalse(id int) (*Cluster, error)
+	FindAllActive() ([]*model.Cluster, error)
+	FindById(id int) (*model.Cluster, error)
+	FindByIdWithActiveFalse(id int) (*model.Cluster, error)
 }
 
-func (impl ClusterRepositoryImpl) FindAllActive() ([]*Cluster, error) {
-	var clusters []*Cluster
+func (impl ClusterRepositoryImpl) FindAllActive() ([]*model.Cluster, error) {
+	var clusters []*model.Cluster
 	err := impl.dbConnection.
 		Model(&clusters).
 		Column("cluster.*", "RemoteConnectionConfig").
@@ -86,8 +56,8 @@ func (impl ClusterRepositoryImpl) FindAllActive() ([]*Cluster, error) {
 	return clusters, err
 }
 
-func (impl ClusterRepositoryImpl) FindById(id int) (*Cluster, error) {
-	var cluster Cluster
+func (impl ClusterRepositoryImpl) FindById(id int) (*model.Cluster, error) {
+	var cluster model.Cluster
 	err := impl.dbConnection.
 		Model(&cluster).
 		Column("cluster.*", "RemoteConnectionConfig").
@@ -97,8 +67,8 @@ func (impl ClusterRepositoryImpl) FindById(id int) (*Cluster, error) {
 	return &cluster, err
 }
 
-func (impl ClusterRepositoryImpl) FindByIdWithActiveFalse(id int) (*Cluster, error) {
-	var cluster Cluster
+func (impl ClusterRepositoryImpl) FindByIdWithActiveFalse(id int) (*model.Cluster, error) {
+	var cluster model.Cluster
 	err := impl.dbConnection.
 		Model(&cluster).
 		Column("cluster.*", "RemoteConnectionConfig").
